+++ app/transforms/file.js
@@ -1,2 +1,0 @@
-import FileTransform from 'ember-flexberry/transforms/file';
-export default FileTransform;