+++ blueprints/flexberry-enum/files/__root__/enums/__name__.js
@@ -1,3 +1,3 @@
-import { createEnum } from 'ember-flexberry/utils/enum-functions';
+import { createEnum } from 'ember-flexberry-data/utils/enum-functions';
 
 export default createEnum(<%= enumObjects %>);