+++ blueprints/flexberry-enum/files/__root__/transforms/__name__.js
@@ -1,4 +1,4 @@
-import FlexberryEnum from 'ember-flexberry/transforms/flexberry-enum';
+import FlexberryEnum from 'ember-flexberry-data/transforms/flexberry-enum';
 import <%= className %>Enum from '../enums/<%= name %>';
 
 export default FlexberryEnum.extend({