+++ tests/dummy/app/transforms/components-examples/flexberry-groupedit/shared/detail-enumeration.js
@@ -1,4 +1,4 @@
-import FlexberryEnumTransormation from 'ember-flexberry/transforms/flexberry-enum';
+import FlexberryEnumTransormation from 'ember-flexberry-data/transforms/flexberry-enum';
 import DetailEnumeration from '../../../../enums/components-examples/flexberry-groupedit/shared/detail-enumeration';
 
 export default FlexberryEnumTransormation.extend({