+++ tests/dummy/app/transforms/ember-flexberry-dummy-gender.js
@@ -1,4 +1,4 @@
-import FlexberryEnum from 'ember-flexberry/transforms/flexberry-enum';
+import FlexberryEnum from 'ember-flexberry-data/transforms/flexberry-enum';
 import GenderEnum from '../enums/ember-flexberry-dummy-gender';
 
 export default FlexberryEnum.extend({