+++ tests/dummy/app/transforms/integration-examples/edit-form/readonly-mode/enumeration.js
@@ -1,4 +1,4 @@
-import FlexberryEnumTransormation from 'ember-flexberry/transforms/flexberry-enum';
+import FlexberryEnumTransormation from 'ember-flexberry-data/transforms/flexberry-enum';
 import Enumeration from '../../../../enums/integration-examples/edit-form/readonly-mode/enumeration';
 
 export default FlexberryEnumTransormation.extend({